Contour CE plane

Contour CE plane

# S3 method for bcea
contour(
  he,
  scale = 0.5,
  nlevels = 4,
  levels = NULL,
  pos = c(1, 0),
  xlim = NULL,
  ylim = NULL,
  graph = c("base", "ggplot2"),
  ...
)

Arguments

he
A bcea object containing the results of the Bayesian modelling and the economic evaluation.

scale
Scales the plot as a function of the observed standard deviation.

nlevels
Number of levels to be plotted in the contour.

levels
Numeric vector of levels at which to draw contour lines. Will be ignored using graph="ggplot2".

pos
Parameter to set the position of the legend (only relevant for multiple interventions, ie more than 2 interventions being compared). Can be given in form of a string (bottom|top)(right|left) for base graphics and bottom|top|left|right for ggplot2. It can be a two-elements vector, which specifies the relative position on the x and y axis respectively, or alternatively it can be in form of a logical variable, with FALSE indicating to use the default position and TRUE to place it on the bottom of the plot.

xlim
The range of the plot along the x-axis. If NULL (default) it is determined by the range of the simulated values for delta_e

ylim
The range of the plot along the y-axis. If NULL (default) it is determined by the range of the simulated values for delta_c

graph
A string used to select the graphical engine to use for plotting. Should (partial-) match the two options "base" or "ggplot2". Default value is "base".

...
Additional arguments to 'plot.window', 'title', 'Axis' and 'box', typically graphical parameters such as 'cex.axis'. Will be ignored if graph="ggplot2".