Specialised contour plot for objects in the class "bcea"

Produces a scatterplot of the cost-effectiveness plane, with a contour-plot of the bivariate density of the differentials of cost (y-axis) and effectiveness (x-axis). Also adds the sustainability area (i.e. below the selected value of the willingness-to-pay threshold).

contour2(
  he,
  wtp = 25000,
  xlim = NULL,
  ylim = NULL,
  comparison = NULL,
  graph = c("base", "ggplot2"),
  ...
)

Arguments

he
A "bcea" object containing the results of the Bayesian modelling and the economic evaluation

wtp
The selected value of the willingness-to-pay. Default is 25000.

xlim
Limits on the x-axis (default=NULL, so that R will select appropriate limits).

ylim
Limits on the y-axis (default=NULL, so that R will select appropriate limits).

comparison
The comparison being plotted. Default to NULL chooses the first comparison if graph="base". If graph="ggplot2" the default value will choose all the possible comparisons. Any subset of the possible comparisons can be selected (e.g., comparison=c(1,3)).

graph
A string used to select the graphical engine to use for plotting. Should (partial-)match the two options "base" or "ggplot2". Default value is "base".

...
Arguments to be passed to ceplane.plot. See the relative manual page for more details.

Value

contour

A ggplot item containing the requested plot. Returned only if graph="ggplot2".

Plots the cost-effectiveness plane with a scatterplot of all the simulated values from the (posterior) bivariate distribution of (Delta_e,Delta_c), the differentials of effectiveness and costs; superimposes a contour of the distribution and prints the value of the ICER, together with the sustainability area.
References

Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity Analysis in Health Economics. Statistical Methods in Medical Research doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London

See also

Examples

### create the bcea object m for the smoking cessation example
data(Smoking)
m=bcea(e,c,ref=4,interventions=treats,Kmax=500)
### produce the plot
contour2(m,wtp=200,graph="base")

#> The first available comparison will be selected.
#>               To plot multiple comparisons together please use the ggplot2 version. Please see ?contour2 for additional details.

#> Error in if (he$n.comparisons > 1) {    if (!exists("title", where = exArgs)) {        title <- paste("Cost effectiveness plane contour plot \n",             he$interventions[he$ref], " vs ", he$interventions[he$comp[comparison]],             sep = "")    }    else {        title <- exArgs$title    }    he$delta.e <- he$delta.e[, comparison]    he$delta.c <- he$delta.c[, comparison]    he$comp <- he$comp[comparison]    he$ICER <- he$ICER[comparison]}: argument is of length zero

# \donttest{
### or use ggplot2 to plot multiple comparisons
contour2(m,wtp=200,ICER.size=2,graph="ggplot2")

#> Error in if (he$n.comparisons == 1) {    density <- with(he, MASS::kde2d(delta.e, delta.c, n = 300,         h = c(sd(delta.e)/scale, sd(delta.c)/scale)))    density <- data.frame(expand.grid(e = density$x, c = density$y),         z = as.vector(density$z))    contour <- ceplane.plot(he, wtp = wtp, graph = "ggplot2",         ...) + ggplot2::geom_contour(ggplot2::aes(z = z, x = e,         y = c), data = density, colour = "black", bins = nlevels)}: argument is of length zero

# }