Expected Value of Information (EVI) plot

Plots the Expected Value of Information (EVI) against the willingness to pay

evi.plot(he, graph = c("base", "ggplot2", "plotly"), ...)

Arguments

he
A bcea object containing the results of the Bayesian modelling and the economic evaluation.

graph
A string used to select the graphical engine to use for plotting. Should (partial-)match the three options "base", "ggplot2" or "plotly". Default value is "base".

...
Additional graphical arguments:

line_colors to specify the EVPI line colour - all graph types.

line_types to specify the line type (lty) - all graph types.

area_include to specify whether to include the area under the EVPI curve - plotly only.

area_color to specify the area under the colour curve - plotly only.

Value

eib

If graph="ggplot2" a ggplot object, or if graph="plotly" a plotly object containing the requested plot. Nothing is returned when graph="base", the default.

The function produces a plot of the Expected Value of Information as a function of the discrete grid approximation of the willingness to pay parameter. The break even point(s) (i.e. the point in which the EIB=0, ie when the optimal decision changes from one intervention to another) is(are) also showed.
References

Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity Analysis in Health Economics. Statistical Methods in Medical Research doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London

See also