Make Report

Constructs the automated report from the output of the BCEA.

Usage

make.report(he, evppi = NULL, ext = "pdf", echo = FALSE, ...)

Arguments

he

A bcea object containing the results of the Bayesian modelling and the economic evaluation.

evppi

An object obtained as output to a call to evppi (default is NULL, so not essential to producing the report).

ext

A string of text to indicate the extension of the resulting output file. Possible options are "pdf", "docx". This requires the use of pandoc, knitr and rmarkdown.

echo

A string (default to FALSE) to instruct whether the report should also include the BCEA commands used to produce the analyses. If the optional argument echo is set to TRUE (default = FALSE), then the commands are also printed.

...

Additional parameters. For example, the user can specify the value of the willingness to pay wtp, which is used in some of the resulting analyses (default at the break even point). Another additional parameter that the user can specify is the name of the file to which the report should be written. This can be done by simply passing the optional argument filename="NAME". The user can also specify an object including the PSA simulations for all the relevant model parameters. If this is passed to the function (in the object psa_sims), then make.report will automatically construct an "Info-rank plot", which is a probabilistic form of tornado plot, based on the Expected Value of Partial Information. The user can also specify the optional argument show.tab (default=FALSE); if set to TRUE, then a table with the values of the Info-rank is also shown.

References

Baio G, Dawid aP (2011). “Probabilistic sensitivity analysis in health economics.” Stat. Methods Med. Res., 1–20. ISSN 1477-0334, doi:10.1177/0962280211419832 , https://pubmed.ncbi.nlm.nih.gov/21930515/.

Baio G (2013). Bayesian Methods in Health Economics. CRC.

See also

Author

Gianluca Baio

Examples

if (FALSE) { # \dontrun{
  data(Vaccine, package = "BCEA")
  m <- bcea(eff, cost, ref = 2)
  make.report(m)
} # }