Summary plot of the health economic analysis when the mixed analysis is considered

Compares the optimal scenario to the mixed case in terms of the EVPI

# S3 method for mixedAn
plot(x, y.limits = NULL, pos = c(0, 1), graph = c("base", "ggplot2"), ...)

Arguments

x
An object of class mixedAn, given as output of the call to the function mixedAn.

y.limits
Range of the y-axis for the graph. The default value is NULL, in which case the maximum range between the optimal and the mixed analysis scenarios is considered.

pos
Parameter to set the position of the legend. Can be given in form of a string (bottom|top)(right|left) for base graphics and bottom|top|left|right for ggplot2. It can be a two-elements vector, which specifies the relative position on the x and y axis respectively, or alternatively it can be in form of a logical variable, with FALSE indicating to use the default position and TRUE to place it on the bottom of the plot. Default value is c(0,1), that is in the topleft corner inside the plot area.

graph
A string used to select the graphical engine to use for plotting. Should (partial-)match the two options "base" or "ggplot2". Default value is "base".

...
Arguments to be passed to methods, such as graphical parameters (see par).

Value

evi

A ggplot object containing the plot. Returned only if graph="ggplot2".

The function produces a graph showing the difference between the ''optimal'' version of the EVPI (when only the most cost-effective intervention is included in the market) and the mixed strategy one (when more than one intervention is considered in the market).
References

Baio, G. and Russo, P. (2009).A decision-theoretic framework for the application of cost-effectiveness analysis in regulatory processes. Pharmacoeconomics 27(8), 645-655 doi:10.2165/11310250

Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity Analysis in Health Economics. Statistical Methods in Medical Research doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London

See also

Examples

# See Baio G., Dawid A.P. (2011) for a detailed description of the 
# Bayesian model and economic problem
#
# Load the processed results of the MCMC simulation model
data(Vaccine)

# Runs the health economic evaluation using BCEA
m <- bcea(e=e,c=c,          # defines the variables of 
                            #  effectiveness and cost
      ref=2,                # selects the 2nd row of (e,c) 
                            #  as containing the reference intervention
      interventions=treats, # defines the labels to be associated 
                            #  with each intervention
      Kmax=50000,           # maximum value possible for the willingness 
                            #  to pay threshold; implies that k is chosen 
                            #  in a grid from the interval (0,Kmax)
      plot=FALSE            # inhibits graphical output
)

ma <- mixedAn(m,        # uses the results of the mixed strategy 
                        #  analysis (a "mixedAn" object)
       mkt.shares=NULL  # the vector of market shares can be defined 
                        #  externally. If NULL, then each of the T 
                        #  interventions will have 1/T market share
)

#> Error in rep(1, he$n.comparators): invalid 'times' argument

# Can also plot the summary graph
plot(ma,graph="base")

#> Error in plot(ma, graph = "base"): object 'ma' not found

# Or with ggplot2
if(require(ggplot2)){
plot(ma,graph="ggplot2")
}

#> Error in plot(ma, graph = "ggplot2"): object 'ma' not found