Table of Simulation Statistics for the Health Economic Model

Using the input in the form of MCMC simulations and after having run the health economic model, produces a summary table of the simulations from the cost-effectiveness analysis.

Usage

sim_table(he, ...)

# S3 method for bcea
sim_table(he, wtp = 25000, ...)

Arguments

he

A bcea object containing the results of the Bayesian modelling and the economic evaluation.

...

Additional arguments

wtp

The value of the willingness to pay threshold to be used in the summary table.

Value

Produces the following elements:

table

A table with simulation statistics from the economic model

names.cols

A vector of labels to be associated with each column of the table

wtp

The selected value of the willingness to pay

idx_wtp

The index associated with the selected value of the willingness to pay threshold in the grid used to run the analysis

References

Baio G, Dawid aP (2011). “Probabilistic sensitivity analysis in health economics.” Stat. Methods Med. Res., 1--20. ISSN 1477-0334, doi:10.1177/0962280211419832 , https://pubmed.ncbi.nlm.nih.gov/21930515/.

Baio G (2013). Bayesian Methods in Health Economics. CRC.

See also

Author

Gianluca Baio

Examples

# See Baio G., Dawid A.P. (2011) for a detailed description of the 
# Bayesian model and economic problem

# Load the processed results of the MCMC simulation model
data(Vaccine)

# Runs the health economic evaluation using BCEA
m <- bcea(e=eff,                # defines the variables of 
          c=cost,               # effectiveness and cost
          ref=2,                # selects the 2nd row of (e, c) 
                                # as containing the reference intervention
          interventions=treats, # defines the labels to be associated 
                                # with each intervention
          Kmax=50000)           # maximum value possible for the willingness 
                                # to pay threshold; implies that k is chosen 
                                # in a grid from the interval (0, Kmax)

# Now can save the simulation exercise in an object using sim_table()
sim_table(m,         # uses the results of the economic evaluation 
          wtp=25000) # selects the particular value for k
#> $Table
#>             U1        U2        U*        IB2_1          OL          VI
#> 1    -36.57582 -38.71760 -36.57582  -2.14178657  2.14178657  -1.7501206
#> 2    -27.92514 -27.67448 -27.67448   0.25065730  0.00000000   7.1512167
#> 3    -28.03024 -33.37394 -28.03024  -5.34369634  5.34369634   6.7954510
#> 4    -53.28408 -47.13734 -47.13734   6.14673837  0.00000000 -12.3116465
#> 5    -43.58389 -40.40469 -40.40469   3.17919761  0.00000000  -5.5789957
#> 6    -42.37456 -33.08547 -33.08547   9.28909873  0.00000000   1.7402301
#> 7    -32.44236 -37.50684 -32.44236  -5.06448157  5.06448157   2.3833392
#> 8    -51.93848 -44.82658 -44.82658   7.11189485  0.00000000 -10.0008876
#> 9    -24.63641 -23.98616 -23.98616   0.65025601  0.00000000  10.8395397
#> 10   -46.18124 -37.34967 -37.34967   8.83156713  0.00000000  -2.5239782
#> 11   -58.09586 -45.57886 -45.57886  12.51699594  0.00000000 -10.7531676
#> 12   -32.02617 -30.02358 -30.02358   2.00258785  0.00000000   4.8021164
#> 13   -24.89058 -26.54063 -24.89058  -1.65005673  1.65005673   9.9351178
#> 14   -51.08339 -34.23447 -34.23447  16.84891677  0.00000000   0.5912225
#> 15   -70.97537 -50.83694 -50.83694  20.13843172  0.00000000 -16.0112433
#> 16   -32.34358 -35.74223 -32.34358  -3.39865390  3.39865390   2.4821190
#> 17   -11.33776 -16.12295 -11.33776  -4.78519176  4.78519176  23.4879376
#> 18   -39.75199 -29.86185 -29.86185   9.89013992  0.00000000   4.9638404
#> 19   -22.20918 -20.04740 -20.04740   2.16178179  0.00000000  14.7782978
#> 20   -45.63434 -62.35286 -45.63434 -16.71851948 16.71851948 -10.8086492
#> 21   -57.54646 -58.46125 -57.54646  -0.91478660  0.91478660 -22.7207650
#> 22   -55.81515 -47.81379 -47.81379   8.00135537  0.00000000 -12.9880959
#> 23   -65.97913 -66.06333 -65.97913  -0.08419258  0.08419258 -31.1534386
#> 24   -15.83557 -26.56841 -15.83557 -10.73283646 10.73283646  18.9901204
#> 25   -21.05875 -35.30491 -21.05875 -14.24616500 14.24616500  13.7669495
#> 26   -60.52026 -42.14622 -42.14622  18.37404259  0.00000000  -7.3205220
#> 27   -28.20554 -35.51466 -28.20554  -7.30912030  7.30912030   6.6201553
#> 28   -32.54679 -35.80048 -32.54679  -3.25368384  3.25368384   2.2789026
#> 29   -16.38009 -21.60875 -16.38009  -5.22865809  5.22865809  18.4456024
#> 30   -13.45785 -14.79877 -13.45785  -1.34092528  1.34092528  21.3678485
#> 31   -65.58315 -52.36537 -52.36537  13.21777897  0.00000000 -17.5396747
#> 32   -17.02895 -19.81625 -17.02895  -2.78729940  2.78729940  17.7967447
#> 33   -26.90749 -24.45174 -24.45174   2.45574173  0.00000000  10.3739514
#> 34   -31.51652 -28.31237 -28.31237   3.20415245  0.00000000   6.5133254
#> 35   -30.04010 -30.98166 -30.04010  -0.94156252  0.94156252   4.7855953
#> 36   -41.82956 -34.43487 -34.43487   7.39469334  0.00000000   0.3908257
#> 37   -47.16277 -40.30695 -40.30695   6.85582892  0.00000000  -5.4812507
#> 38   -23.39792 -36.38016 -23.39792 -12.98223797 12.98223797  11.4277781
#> 39   -45.96626 -42.40697 -42.40697   3.55929869  0.00000000  -7.5812700
#> 40   -30.72831 -33.35439 -30.72831  -2.62607984  2.62607984   4.0973832
#> 41   -50.29058 -46.21712 -46.21712   4.07346281  0.00000000 -11.3914237
#> 42   -21.74774 -29.78883 -21.74774  -8.04108588  8.04108588  13.0779516
#> 43   -36.74093 -30.35715 -30.35715   6.38378160  0.00000000   4.4685494
#> 44   -24.90551 -39.28508 -24.90551 -14.37957453 14.37957453   9.9201871
#> 45   -53.62945 -39.94290 -39.94290  13.68654669  0.00000000  -5.1172077
#> 46   -32.73120 -28.45269 -28.45269   4.27851070  0.00000000   6.3730062
#> 47   -38.81344 -37.71039 -37.71039   1.10305087  0.00000000  -2.8846976
#> 48   -60.41305 -46.71309 -46.71309  13.69996697  0.00000000 -11.8873899
#> 49   -21.07820 -19.70921 -19.70921   1.36898540  0.00000000  15.1164839
#> 50   -31.28313 -32.32021 -31.28313  -1.03708020  1.03708020   3.5425639
#> 51   -32.45005 -41.29502 -32.45005  -8.84497820  8.84497820   2.3756487
#> 52   -53.77356 -50.06956 -50.06956   3.70399842  0.00000000 -15.2438679
#> 53   -39.25631 -31.63766 -31.63766   7.61865192  0.00000000   3.1880324
#> 54   -32.09595 -28.83281 -28.83281   3.26313571  0.00000000   5.9928847
#> 55   -41.25195 -44.26889 -41.25195  -3.01694505  3.01694505  -6.4262503
#> 56   -56.13018 -46.32287 -46.32287   9.80730420  0.00000000 -11.4971758
#> 57   -23.75143 -41.63394 -23.75143 -17.88250919 17.88250919  11.0742605
#> 58   -36.16824 -29.83677 -29.83677   6.33146854  0.00000000   4.9889249
#> 59   -31.31161 -31.06406 -31.06406   0.24755007  0.00000000   3.7616355
#> 60   -57.96474 -44.05080 -44.05080  13.91394349  0.00000000  -9.2251021
#> 61   -42.58116 -36.25501 -36.25501   6.32614887  0.00000000  -1.4293166
#> 62   -39.89875 -29.85403 -29.85403  10.04472122  0.00000000   4.9716621
#> 63   -53.78104 -44.48163 -44.48163   9.29940909  0.00000000  -9.6559357
#> 64   -69.61427 -48.36707 -48.36707  21.24720269  0.00000000 -13.5413727
#> 65   -16.46813 -25.18588 -16.46813  -8.71774439  8.71774439  18.3575641
#> 66   -19.60455 -26.60639 -19.60455  -7.00183629  7.00183629  15.2211441
#> 67   -45.40540 -43.17445 -43.17445   2.23095557  0.00000000  -8.3487506
#> 68   -26.86451 -27.54852 -26.86451  -0.68401927  0.68401927   7.9611902
#> 69   -37.60752 -36.44240 -36.44240   1.16511688  0.00000000  -1.6167055
#> 70   -38.48746 -41.94032 -38.48746  -3.45286217  3.45286217  -3.6617659
#> 71   -23.02737 -25.09422 -23.02737  -2.06685854  2.06685854  11.7983302
#> 72   -14.33656 -18.17653 -14.33656  -3.83996881  3.83996881  20.4891371
#> 73   -45.76326 -37.42407 -37.42407   8.33918599  0.00000000  -2.5983748
#> 74   -38.29317 -38.54554 -38.29317  -0.25237311  0.25237311  -3.4674737
#> 75   -50.17215 -47.85474 -47.85474   2.31741201  0.00000000 -13.0290403
#> 76   -23.80987 -22.01309 -22.01309   1.79677716  0.00000000  12.8126024
#> 77   -37.74850 -30.80923 -30.80923   6.93927189  0.00000000   4.0164658
#> 78   -23.52495 -27.07055 -23.52495  -3.54559993  3.54559993  11.3007412
#> 79   -47.14022 -32.83611 -32.83611  14.30410435  0.00000000   1.9895844
#> 80   -25.06311 -24.82709 -24.82709   0.23601979  0.00000000   9.9986099
#> 81   -37.67666 -35.71838 -35.71838   1.95827838  0.00000000  -0.8926846
#> 82   -27.96153 -37.20587 -27.96153  -9.24434416  9.24434416   6.8641679
#> 83   -26.48804 -32.34358 -26.48804  -5.85553716  5.85553716   8.3376568
#> 84   -17.81888 -28.87161 -17.81888 -11.05272772 11.05272772  17.0068125
#> 85   -39.00951 -35.86738 -35.86738   3.14213033  0.00000000  -1.0416826
#> 86   -38.53378 -31.32709 -31.32709   7.20669147  0.00000000   3.4986026
#> 87   -36.48633 -27.86485 -27.86485   8.62148211  0.00000000   6.9608469
#> 88   -59.78943 -57.11228 -57.11228   2.67714547  0.00000000 -22.2865892
#> 89   -24.15185 -26.45561 -24.15185  -2.30375471  2.30375471  10.6738423
#> 90   -16.37085 -19.69877 -16.37085  -3.32792197  3.32792197  18.4548482
#> 91   -29.63276 -29.51358 -29.51358   0.11917430  0.00000000   5.3121116
#> 92   -39.89357 -39.49942 -39.49942   0.39414801  0.00000000  -4.6737253
#> 93   -21.27230 -33.24018 -21.27230 -11.96788087 11.96788087  13.5533970
#> 94   -30.38773 -37.05681 -30.38773  -6.66908550  6.66908550   4.4379703
#> 95   -33.35082 -31.41187 -31.41187   1.93894462  0.00000000   3.4138238
#> 96   -40.02950 -35.15516 -35.15516   4.87434345  0.00000000  -0.3294661
#> 97   -17.70239 -26.73252 -17.70239  -9.03012465  9.03012465  17.1233025
#> 98   -16.12799 -24.72537 -16.12799  -8.59738436  8.59738436  18.6977076
#> 99   -50.58139 -43.91244 -43.91244   6.66894479  0.00000000  -9.0867474
#> 100  -64.22113 -53.02862 -53.02862  11.19251029  0.00000000 -18.2029217
#> 101  -36.13026 -35.79804 -35.79804   0.33221756  0.00000000  -0.9723451
#> 102  -35.36210 -34.25108 -34.25108   1.11102260  0.00000000   0.5746170
#> 103  -35.29424 -30.87344 -30.87344   4.42080134  0.00000000   3.9522521
#> 104  -57.27238 -38.28796 -38.28796  18.98441950  0.00000000  -3.4622602
#> 105  -32.47935 -25.94254 -25.94254   6.53681683  0.00000000   8.8831597
#> 106  -39.64396 -32.67332 -32.67332   6.97064482  0.00000000   2.1523766
#> 107  -31.19543 -30.59790 -30.59790   0.59753035  0.00000000   4.2277907
#> 108  -75.36875 -68.87114 -68.87114   6.49760801  0.00000000 -34.0454456
#> 109  -61.75084 -40.70930 -40.70930  21.04153963  0.00000000  -5.8836020
#> 110  -20.31745 -29.21184 -20.31745  -8.89438758  8.89438758  14.5082437
#> 111  -27.51296 -33.35347 -27.51296  -5.84050906  5.84050906   7.3127313
#> 112  -36.48112 -33.36829 -33.36829   3.11282481  0.00000000   1.4574049
#> 113  -32.82014 -38.43081 -32.82014  -5.61066138  5.61066138   2.0055504
#> 114  -36.54572 -41.35659 -36.54572  -4.81087447  4.81087447  -1.7200248
#> 115  -32.68648 -32.78388 -32.68648  -0.09740091  0.09740091   2.1392113
#> 116  -23.51571 -25.64790 -23.51571  -2.13218539  2.13218539  11.3099849
#> 117  -25.26737 -26.91231 -25.26737  -1.64494678  1.64494678   9.5583275
#> 118  -37.76828 -36.61777 -36.61777   1.15051016  0.00000000  -1.7920719
#> 119  -26.37334 -25.00879 -25.00879   1.36455038  0.00000000   9.8169037
#> 120  -29.41048 -27.68329 -27.68329   1.72719676  0.00000000   7.1424091
#> 121  -18.10175 -21.30367 -18.10175  -3.20192350  3.20192350  16.7239473
#> 122  -38.64128 -32.76287 -32.76287   5.87840807  0.00000000   2.0628204
#> 123  -33.76574 -44.97225 -33.76574 -11.20650432 11.20650432   1.0599525
#> 124  -23.77004 -32.16060 -23.77004  -8.39056052  8.39056052  11.0556510
#> 125  -30.70120 -40.14649 -30.70120  -9.44528345  9.44528345   4.1244906
#> 126  -50.58272 -46.21238 -46.21238   4.37033951  0.00000000 -11.3866836
#> 127  -20.97638 -22.33520 -20.97638  -1.35882724  1.35882724  13.8493179
#> 128  -22.30573 -22.21149 -22.21149   0.09424148  0.00000000  12.6142084
#> 129  -21.09558 -29.42030 -21.09558  -8.32472146  8.32472146  13.7301180
#> 130  -73.16543 -60.35362 -60.35362  12.81181608  0.00000000 -25.5279224
#> 131  -14.16781 -17.44756 -14.16781  -3.27974739  3.27974739  20.6578873
#> 132  -35.89800 -27.55970 -27.55970   8.33829959  0.00000000   7.2659925
#> 133  -40.27088 -39.93264 -39.93264   0.33824044  0.00000000  -5.1069478
#> 134  -38.73472 -41.17318 -38.73472  -2.43846645  2.43846645  -3.9090214
#> 135  -29.15612 -30.85489 -29.15612  -1.69876977  1.69876977   5.6695789
#> 136  -22.95564 -21.33305 -21.33305   1.62258412  0.00000000  13.4926428
#> 137  -62.25290 -42.70410 -42.70410  19.54879905  0.00000000  -7.8784053
#> 138 -102.17776 -98.97074 -98.97074   3.20702071  0.00000000 -64.1450423
#> 139  -32.79665 -32.64367 -32.64367   0.15298190  0.00000000   2.1820283
#> 140  -47.67811 -33.08435 -33.08435  14.59375655  0.00000000   1.7413422
#> 141  -21.15178 -25.22621 -21.15178  -4.07443816  4.07443816  13.6739188
#> 142  -32.12205 -30.08375 -30.08375   2.03829819  0.00000000   4.7419433
#> 143  -26.29690 -21.54334 -21.54334   4.75356647  0.00000000  13.2823598
#> 144  -43.02542 -36.88249 -36.88249   6.14292922  0.00000000  -2.0567982
#> 145  -42.77674 -36.51205 -36.51205   6.26469185  0.00000000  -1.6863516
#> 146  -37.56970 -36.68284 -36.68284   0.88685685  0.00000000  -1.8571496
#> 147  -24.20949 -27.40378 -24.20949  -3.19429253  3.19429253  10.6162095
#> 148  -47.98224 -48.38364 -47.98224  -0.40139562  0.40139562 -13.1565460
#> 149  -77.71323 -40.99638 -40.99638  36.71684922  0.00000000  -6.1706838
#> 150  -43.76933 -42.28314 -42.28314   1.48619442  0.00000000  -7.4574406
#> 151  -34.69471 -34.31585 -34.31585   0.37885997  0.00000000   0.5098435
#> 152  -41.45332 -36.63997 -36.63997   4.81334506  0.00000000  -1.8142778
#> 153  -16.00529 -31.02543 -16.00529 -15.02013948 15.02013948  18.8204068
#> 154  -25.71344 -28.85638 -25.71344  -3.14293739  3.14293739   9.1122516
#> 155  -37.22640 -35.77600 -35.77600   1.45039615  0.00000000  -0.9503060
#> 156  -55.14584 -48.32931 -48.32931   6.81652031  0.00000000 -13.5036196
#> 157  -32.06361 -34.97150 -32.06361  -2.90788877  2.90788877   2.7620843
#> 158  -26.91445 -27.30613 -26.91445  -0.39167901  0.39167901   7.9112480
#> 159  -30.46535 -34.34856 -30.46535  -3.88320810  3.88320810   4.3603411
#> 160  -17.12845 -22.01007 -17.12845  -4.88162304  4.88162304  17.6972497
#> 161  -18.98263 -21.73387 -18.98263  -2.75124107  2.75124107  15.8430634
#> 162  -29.62740 -31.66658 -29.62740  -2.03918727  2.03918727   5.1983000
#> 163  -48.90434 -46.09079 -46.09079   2.81355383  0.00000000 -11.2650914
#> 164  -18.96137 -37.81434 -18.96137 -18.85297646 18.85297646  15.8643294
#> 165  -54.05352 -40.70611 -40.70611  13.34740946  0.00000000  -5.8804113
#> 166  -23.36848 -27.95576 -23.36848  -4.58727421  4.58727421  11.4572107
#>  [ reached 'max' / getOption("max.print") -- omitted 835 rows ]
#> 
#> $names.cols
#> [1] "U1"    "U2"    "U*"    "IB2_1" "OL"    "VI"   
#> 
#> $wtp
#> [1] 25000
#> 
#> $ind.table
#> [1] 251
#>