Structural Probability Sensitivity Analysis

Computes the weights to be associated with a set of competing models in order to perform structural PSA.

struct.psa(
  models,
  effect,
  cost,
  ref = NULL,
  interventions = NULL,
  Kmax = 50000,
  plot = FALSE
)

Arguments

models
A list containing the output from either R2jags or R2OpenBUGS/R2WinBUGS for all the models that need to be combined in the model average

effect
A list containing the measure of effectiveness computed from the various models (one matrix with n.sim x n.ints simulations for each model)

cost
A list containing the measure of costs computed from the various models (one matrix with n.sim x n.ints simulations for each model)

ref
Which intervention is considered to be the reference strategy. The default value ref=1 means that the intervention appearing first is the reference and the other(s) is(are) the comparator(s)

interventions
Defines the labels to be associated with each intervention. By default and if NULL, assigns labels in the form "Intervention1", ... , "InterventionT"

Kmax
Maximum value of the willingness to pay to be considered. Default value is k=50000. The willingness to pay is then approximated on a discrete grid in the interval [0,Kmax]. The grid is equal to wtp if the parameter is given, or composed of 501 elements if wtp=NULL (the default)

plot
A logical value indicating whether the function should produce the summary plot or not

Value

List object of bcea object, model weights and DIC

Details

The model is a list containing the output from either R2jags or R2OpenBUGS / R2WinBUGS for all the models that need to be combined in the model average effect is a list containing the measure of effectiveness computed from the various models (one matrix with n_sim x n_ints simulations for each model) cost is a list containing the measure of costs computed from the various models (one matrix with n_sim x n_ints simulations for each model).

References

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London.

See also

Author

Gianluca Baio