Summary Method for Objects of Class bcea

Produces a table printout with some summary results of the health economic evaluation.

Usage

# S3 method for class 'bcea'
summary(object, wtp = 25000, ...)

Arguments

object

A bcea object containing the results of the Bayesian modelling and the economic evaluation.

wtp

The value of the willingness to pay threshold used in the summary table.

...

Additional arguments affecting the summary produced.

Value

Prints a summary table with some information on the health economic output and synthetic information on the economic measures (EIB, CEAC, EVPI).

References

Baio2011BCEA

Baio2013BCEA

See also

Author

Gianluca Baio

Examples

data(Vaccine)

he <- bcea(eff, cost, interventions = treats, ref = 2)
summary(he)
#> 
#> Cost-effectiveness analysis summary 
#> 
#> Reference intervention:  Vaccination
#> Comparator intervention: Status Quo
#> 
#> Optimal decision: choose Status Quo for k < 20100 and Vaccination for k >= 20100
#> 
#> 
#> Analysis for willingness to pay parameter k = 25000
#> 
#>             Expected net benefit
#> Status Quo               -36.054
#> Vaccination              -34.826
#> 
#>                              EIB  CEAC  ICER
#> Vaccination vs Status Quo 1.2284 0.529 20098
#> 
#> Optimal intervention (max expected net benefit) for k = 25000: Vaccination
#>            
#> EVPI 2.4145