Data set for the Bayesian model for the cost-effectiveness of smoking cessation interventions

This data set contains the results of the Bayesian analysis used to model the clinical output and the costs associated with the health economic evaluation of four different smoking cessation interventions.

Format

A data list including the variables needed for the smoking cessation cost-effectiveness analysis. The variables are as follows:

list("c")

a matrix of 500 simulations from the posterior distribution of the overall costs associated with the four strategies

list("data")

a dataset containing the characteristics of the smokers in the UK population

list("e")

a matrix of 500 simulations from the posterior distribution of the clinical benefits associated with the four strategies

list("life.years")

a matrix of 500 simulations from the posterior distribution of the life years gained with each strategy

list("pi")

a matrix of 500 simulations from the posterior distribution of the event of smoking cessation with each strategy

list("smoking")

a data frame containing the inputs needed for the network meta-analysis model. The data.frame object contains: nobs: the record ID number, s: the study ID number, i: the intervention ID number, r_i: the number of patients who quit smoking, n_i: the total number of patients for the row-specific arm and b_i: the reference intervention for each study

list("smoking_output")

a rjags object obtained by running the network meta-analysis model based on the data contained in the smoking object

list("smoking_mat")

a matrix obtained by running the network meta-analysis model based on the data contained in the smoking object

list("treats")

a vector of labels associated with the four strategies

Source

Effectiveness data adapted from Hasselblad V. (1998). Meta-analysis of Multitreatment Studies. Medical Decision Making 1998;18:37-43.

Cost and population characteristics data adapted from various sources:

Taylor, D.H. Jr, et al. (2002). Benefits of smoking cessation on longevity. American Journal of Public Health 2002;92(6)

ASH: Action on Smoking and Health (2013). ASH fact sheet on smoking statistics,
http://ash.org.uk/files/documents/ASH_106.pdf

Flack, S., et al. (2007). Cost-effectiveness of interventions for smoking cessation. York Health Economics Consortium, January 2007

McGhan, W.F.D., and Smith, M. (1996). Pharmacoeconomic analysis of smoking-cessation interventions. American Journal of Health-System Pharmacy 1996;53:45-52

References

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London

Examples

data(Smoking)

# \donttest{
m=bcea(e,c,ref=4,interventions=treats,Kmax=500)
# }