Data set for the Bayesian model for the cost-effectiveness of influenza vaccination

This data set contains the results of the Bayesian analysis used to model the clinical output and the costs associated with an influenza vaccination.

Format

A data list including the variables needed for the influenza vaccination. The variables are as follows:

list("c")

a matrix of simulations from the posterior distribution of the overall costs associated with the two treatments

list("c.pts")

list("cost.GP")

a matrix of simulations from the posterior distribution of the costs for GP visits associated with the two treatments

list("cost.hosp")

a matrix of simulations from the posterior distribution of the costs for hospitalisations associated with the two treatments

list("cost.otc")

a matrix of simulations from the posterior distribution of the costs for over-the-counter medications associated with the two treatments

list("cost.time.off")

a matrix of simulations from the posterior distribution of the costs for time off work associated with the two treatments

list("cost.time.vac")

a matrix of simulations from the posterior distribution of the costs for time needed to get the vaccination associated with the two treatments

list("cost.travel")

a matrix of simulations from the posterior distribution of the costs for travel to get vaccination associated with the two treatments

list("cost.trt1")

a matrix of simulations from the posterior distribution of the overall costs for first line of treatment associated with the two interventions

list("cost.trt2")

a matrix of simulations from the posterior distribution of the overall costs for second line of treatment associated with the two interventions

list("cost.vac")

a matrix of simulations from the posterior distribution of the costs for vaccination

list("e")

a matrix of simulations from the posterior distribution of the clinical benefits associated with the two treatments

list("e.pts")

list("N")

the number of subjects in the reference population

list("N.outcomes")

the number of clinical outcomes analysed

list("N.resources")

the number of health-care resources under study

list("QALYs.adv")

a vector from the posterior distribution of the QALYs associated with advert events

list("QALYs.death")

a vector from the posterior distribution of the QALYs associated with death

list("QALYs.hosp")

a vector from the posterior distribution of the QALYs associated with hospitalisation

list("QALYs.inf")

a vector from the posterior distribution of the QALYs associated with influenza infection

list("QALYs.pne")

a vector from the posterior distribution of the QALYs associated with pneumonia

list("treats")

a vector of labels associated with the two treatments

list("vaccine")

a rjags object containing the simulations for the parameters used in the original model

list("vaccine_mat")

a matrix containing the simulations for the parameters used in the original model

Source

Adapted from Turner D, Wailoo A, Cooper N, Sutton A, Abrams K, Nicholson K. The cost-effectiveness of influenza vaccination of healthy adults 50-64 years of age. Vaccine. 2006;24:1035-1043.

References

Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity Analysis in Health Economics. Statistical Methods in Medical Research doi:10.1177/0962280211419832.

Examples

data(Vaccine)

# \donttest{
m <- bcea(e, c, ref = 1, interventions = treats)
# }